\name{print.cigraph}
\alias{print.cigraph}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print edgelist
}
\description{
This function prints the edgelist of a "cigraph" object
}
\usage{
\method{print}{cigraph}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
A "cigraph" object
}
\item{...}{
These arguments are not used
}
}

\author{
Sacha Epskamp (mail@sachaepskamp.com)
}

\seealso{
\code{\link{cigraph}}
}
