\name{summary.cigraph}
\alias{summary.cigraph}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Summary method for "cigraph"
}
\description{
This function creates a brief summary based on a "cigraph" object.
}
\usage{
\method{summary}{cigraph}(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
A "cigraph" object
}
\item{...}{
These arguments are not used
}
}

\author{
Sacha Epskamp (mail@sachaepskamp.com)
}

\seealso{
\code{\link{cigraph}}
}
